requirements/base.txt
djangorestframework==3.15.1
Django==4.2.13
django-filter
djangorestframework-simplejwt
djangorestframework-recursive
django-cors-headers
django-redis
redis
tornado
flower
django-redis
redis
tornado
celery
requests
twilio
whatsapp-api-client-python
pywhatkit